+++ BlinkID/src/ios/MicroblinkModule/MicroblinkModule/MBBlinkIDSerializationUtils.h
@@ -31,4 +31,5 @@
 +(NSNumber * _Nullable)serializeMBSide:(MBSide) value;
 +(NSDictionary * _Nonnull) serializeNSDate:(NSDate * _Nullable) value;
 +(MBCustomClassRules * _Nonnull) deserializeMBCustomClassRules:(NSDictionary * _Nullable)jsonCustomClassRules;
++(NSMutableArray<NSDictionary *> * _Nullable) serializeDependentInfo:(NSArray<MBDependentInfo *>* _Nullable)dependentInfos;
 @end